Revert "All files commited"
This reverts commit 104376047cfd474b1884c99576d2fd93458ab8a7.

+++ src/main/java/com/example/eticaret/service/ProductService.java
@@ -23,12 +23,21 @@ public class ProductService {
     public Product findById(Long id) {
         return productRepository.findById(id).orElse(null);
     }
-    public Product addProduct(Product product) {
+    public Product addProduct(Product product,Long userId) {
+        User user =userRepository.findById(userId).orElse(null);
+
+       if(!user.isAdmin()) {
+           return null;
+       }
         return productRepository.save(product);
     }
     public void deleteProduct(Long productId,Long userId) {
         User user =userRepository.findById(userId).orElse(null);
+
+        if(!user.isAdmin()){
+            return;
+        }
             Product product = productRepository.findById(productId).orElse(null);
             productRepository.delete(product);
-    }
+        }
 }
